+++ CHANGELOG.md
@@ -7,6 +7,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 - Added unit test for whitespace handling.
 - Added optional argument `$trim = true` to `$node->text()`
+- Added attribute value normalization
+  - https://www.w3.org/TR/html/syntax.html#attribute-values
+  - https://www.w3.org/TR/xml/#AVNormalize
 ### Changed
 - `$node->text()` now supports all block and inline level elements.
 ### Deprecated